Refactoring #1
These changes are all pure refactoring (no functional changes).
Classes split up;
Name changes, eg. command->flag; and PosixArgParser->Parser
