+++ src/main/scala/com/phasmidsoftware/args/Args.scala
@@ -8,12 +8,12 @@ import com.phasmidsoftware.util.MonadOps._
 import com.phasmidsoftware.util.{Kleenean, Maybe}
 
 import scala.util._
-import scala.util.parsing.combinator.RegexParsers
 
 /**
-  * Case class to represent an "option" in a command line.
+  * Case class to represent an "option" in a flag line.
   * Such an option has an (optional) name which is a String;
   * and an (optional) value, which is of type X.
+  *
   * @param name the optional name.
   * @param value the optional value.
   * @tparam X the underlying type of the value.
@@ -116,9 +116,10 @@ case class Arg[X](name: Option[String], value: Option[X]) extends Ordered[Arg[X]
 
   /**
     * Method to form a String from this Arg
-    * @return "Arg: command name/anonymous with value: value/none"
+    *
+    * @return "Arg: flag name/anonymous with value: value/none"
     */
-  override def toString: String = s"Arg: command ${name.getOrElse("anonymous")} with value: ${value.getOrElse("none")}"
+  override def toString: String = s"Arg: flag ${name.getOrElse("anonymous")} with value: ${value.getOrElse("none")}"
 
   def compare(that: Arg[X]): Int = name match {
     case Some(x) => that.name match {
@@ -150,10 +151,11 @@ case class Args[X](xas: Seq[Arg[X]]) extends Iterable[Arg[X]] {
 
   /**
     * Method to validate this Args according to the POSIX-style synopsis w, expressed as a String.
+    *
     * @param w the synopsis
     * @return this Args, assuming that all is OK
     */
-  def validate(w: String): Args[X] = validate(new PosixSynopsisParser().parseSynopsis(Some(w)))
+  def validate(w: String): Args[X] = validate(new SynopsisParser().parseSynopsis(Some(w)))
 
   /**
     * Method to validate this Args according to the (optional) synopsis given as so.
@@ -185,16 +187,6 @@ case class Args[X](xas: Seq[Arg[X]]) extends Iterable[Arg[X]] {
     else
       false
   }
-
-  /**
-    * Validate this Args with respect to just one Element of a Synopsis.
-    *
-    * TODO: figure out what this is or was supposed to do -- it doesn't get called at present.
-    *
-    * @param e the element
-    * @return false
-    */
-  def validate(e: Element): Boolean = false
 
   /**
     * Apply the given function f to each Arg of this Args
@@ -242,7 +234,8 @@ case class Args[X](xas: Seq[Arg[X]]) extends Iterable[Arg[X]] {
 
   /**
     * Method to determine if the argument identified by w is defined.
-    * @param w the name of an argument (command).
+    *
+    * @param w the name of an argument (flag).
     * @return true if the argument is found by getArg
     */
   def isDefined(w: String): Boolean = getArg(w).isDefined
@@ -275,8 +268,8 @@ object Args {
 
     def inner(r: Seq[Arg[String]], w: Seq[p.Token]): Seq[Arg[String]] = w match {
       case Nil => r
-      case p.Command(c) :: p.Argument(a) :: t => inner(r :+ Arg(c, a), t)
-      case p.Command(c) :: t => inner(r :+ Arg(c), t)
+      case p.Flag(c) :: p.Argument(a) :: t => inner(r :+ Arg(c, a), t)
+      case p.Flag(c) :: t => inner(r :+ Arg(c), t)
       case p.Argument(a) :: t => inner(r :+ Arg(None, Some(a)), t)
     }
 
@@ -290,11 +283,12 @@ object Args {
 
   /**
     * Method to parse a set of command line arguments that conform to the POSIX standard.
+    *
     * @param args the command line arguments.
     * @param synopsis the (optional) syntax template which will be used, if not None, to validate the options.
     * @return the arguments parsed as an Args[String].
     */
-  def parsePosix(args: Array[String], synopsis: Option[String] = None): Args[String] = doParse((new PosixArgParser).parseCommandLine(args), synopsis)
+  def parsePosix(args: Array[String], synopsis: Option[String] = None): Args[String] = doParse((new Parser).parseCommandLine(args), synopsis)
 
   /**
     * Method to create an Args object from a variable number of Arg parameters.
@@ -307,7 +301,7 @@ object Args {
   def create(args: Arg[String]*): Args[String] = apply(args)
 
   private def doParse(ps: Seq[PosixArg], wo: Option[String] = None): Args[String] = {
-    val so = (new PosixSynopsisParser).parseSynopsis(wo)
+    val so = (new SynopsisParser).parseSynopsis(wo)
 
     def processPosixArg(p: PosixArg): Seq[PosixArg] = p match {
       case PosixOptionString(w) =>
@@ -322,7 +316,7 @@ object Args {
                   case Some(e) =>
                     def processElement(e: Element): Seq[PosixArg] = e match {
                       case OptionalElement(x) => processElement(x)
-                      case CommandWithValue(_, OptionalElement(_)) => ws ++ Seq(PosixOptionString(c.toString), PosixOptionValue(tail.mkString("")))
+                      case FlagWithValue(_, OptionalElement(_)) => ws ++ Seq(PosixOptionString(c.toString), PosixOptionValue(tail.mkString("")))
                       case _ => inner2(ws :+ PosixOptionString(c.toString), tail)
                     }
 
@@ -353,290 +347,3 @@ object Args {
   }
 
 }
-
-/**
-  * Type-class trait to allow conversion from type X to type T
-  *
-  * @tparam T the result type
-  */
-trait Derivable[T] {
-  /**
-    * Method to convert an X to a T
-    *
-    * @param x the X value
-    * @tparam X the input type
-    * @return a T
-    */
-  def deriveFrom[X](x: X): T
-}
-
-/**
-  * Parser of non-POSIX command lines
-  */
-class SimpleArgParser extends RegexParsers {
-  def parseToken(s: String): Try[Token] = parseAll(token, s) match {
-    case Success(t, _) => scala.util.Success(t)
-    case _ => scala.util.Failure(new Exception(s"could not parse '$s' as a token"))
-  }
-
-  trait Token {
-    def s: String
-  }
-
-  case class Command(s: String) extends Token
-
-  case class Argument(s: String) extends Token
-
-  def token: Parser[Token] = command | argument
-
-  def command: Parser[Command] = "-" ~> cmdR ^^ (s => Command(s))
-
-  def argument: Parser[Argument] = argR ^^ (s => Argument(s))
-
-  private val cmdR = """[a-z]+""".r
-  private val argR = """\w+""".r
-}
-
-/**
-  * a Posix Arg
-  */
-trait PosixArg {
-  def value: String
-}
-
-/**
-  * One or more options.
-  * Each option is a single-letter, although the terminating
-  * characters can be a value.
-  *
-  * @param value the string of options, without the "-" prefix.
-  */
-case class PosixOptionString(value: String) extends PosixArg
-
-/**
-  * The value of the preceding option.
-  *
-  * @param value a String
-  */
-case class PosixOptionValue(value: String) extends PosixArg
-
-/**
-  * The value of an operand, i.e. a String which follows all of the options and their values.
-  *
-  * @param value a String
-  */
-case class PosixOperand(value: String) extends PosixArg
-
-/**
-  * Parser of POSIX-style command lines.
-  */
-class PosixArgParser extends RegexParsers {
-
-  def parseCommandLine(ws: Seq[String]): Seq[PosixArg] = parseAll(posixCommandLine, ws.mkString("", terminator, terminator)) match {
-    case Success(t, _) => t
-    case _ => throw ParseException(s"could not parse '$ws' as a token")
-  }
-
-  /**
-    * NOTE that it is impossible to tell whether the first arg after an option set is an option value or the first operand.
-    * Only validating it with a command line synopsis can do that for sure.
-    *
-    * @return
-    */
-  def posixCommandLine: Parser[Seq[PosixArg]] = rep(posixOptionSet) ~ rep(posixOperand) ^^ { case pss ~ ps => pss.flatten ++ ps }
-
-  def posixOptionSet: Parser[Seq[PosixArg]] = posixOptions ~ opt(posixOptionValue) ^^ { case p ~ po => p +: po.toSeq }
-
-  def posixOptions: Parser[PosixArg] = "-" ~> """[a-zA-Z0-9]+""".r <~ terminator ^^ (s => PosixOptionString(s))
-
-  def posixOptionValue: Parser[PosixArg] = nonOption ^^ (s => PosixOptionValue(s))
-
-  def posixOperand: Parser[PosixArg] = nonOption ^^ (s => PosixOperand(s))
-
-  def nonOption: Parser[String] = """[^;]+""".r <~ terminator
-
-  val terminator = ";"
-}
-
-/**
-  * This represents an element in the synopsis for a command line
-  */
-trait Element extends Ordered[Element] {
-  def value: String
-
-  def isOptional: Boolean = false
-
-  def compare(that: Element): Int = value compare that.value
-}
-
-/**
-  * NOTE: Don't think we need this
-  */
-trait Optional
-
-case class Synopsis(es: Seq[Element]) {
-  def getElement(w: String): Option[Element] = find(Some(w))
-
-  def find(wo: Option[String]): Option[Element] = wo match {
-    case Some(w) => es.find(e => e.value == w)
-    case _ => None
-  }
-
-  def mandatoryAndOptionalElements: (Seq[Element], Seq[Element]) = es partition (!_.isOptional)
-}
-
-/**
-  * This represents an "Option" in the parlance of POSIX command line interpretation
-  *
-  * CONSIDER: rename this to "Flag"
-  *
-  * @param value the (single-character) String representing the command
-  */
-case class Command(value: String) extends Element
-
-/**
-  * This represents an Option Value in the parlance of POSIX.
-  *
-  * @param value the String
-  */
-case class Value(value: String) extends Element
-
-/**
-  * This represents an "Option" and its "Value"
-  *
-  * @param value   the command or "option" String
-  * @param element the Element which corresponds to the "value" of this synopsis command (and which may of course be OptionalElement).
-  */
-case class CommandWithValue(value: String, element: Element) extends Element {
-  override def equals(obj: scala.Any): Boolean = obj match {
-    case CommandWithValue(x, y) => value == x && element == y
-    case _ => false
-  }
-}
-
-/**
-  * This represents an optional synopsis element, either an optional command, or an optional value.
-  *
-  * @param element a synopsis element that is optional
-  */
-case class OptionalElement(element: Element) extends Element with Optional {
-  def value: String = element.value
-
-  override def isOptional: Boolean = true
-
-  override def equals(obj: scala.Any): Boolean = obj match {
-    case e: Element => e.isOptional && value == e.value
-    case _ => false
-  }
-}
-
-class PosixSynopsisParser extends RegexParsers {
-  def parseSynopsis(wo: Option[String]): Option[Synopsis] = wo match {
-    case Some(w) => parseAll(synopsis, w) match {
-      case Success(es, _) => Some(Synopsis(es))
-      case _ => throw new Exception(s"could not parse '$w' as a synopsis")
-    }
-    case _ => None
-  }
-
-  override def skipWhitespace: Boolean = false
-
-  /**
-    * A "synopsis" of command line options and their potential argument values.
-    * It matches a dash ('-') followed by a list of optionalOrRequiredElement OR: an optional list of optionWithOrWithoutValue
-    *
-    * @return a Parser[Seq[Element]
-    */
-  def synopsis: Parser[Seq[Element]] = "-" ~> (optionalElements | rep(optionalOrRequiredElement))
-
-  /**
-    * An optionalOrRequiredElement matches EITHER: an optionalElement OR: an optionWithOrWithoutValue
-    *
-    * @return a Parser[Element] which is EITHER: a Parser[Command] OR: a Parser[CommandWithValue] OR: a Parser[OptionalElement]
-    */
-  def optionalOrRequiredElement: Parser[Element] = optionalElement | optionWithOrWithoutValue
-
-  /**
-    * An optionalElement matches a '[' followed by an optionWithOrWithoutValue followed by a ']'
-    *
-    * @return a Parser[OptionalElement]
-    */
-  def optionalElement: Parser[Element] = openBracket ~> optionWithOrWithoutValue <~ closeBracket ^^ (t => OptionalElement(t))
-
-  /**
-    * An optionalElement matches a '[' followed by an optionWithOrWithoutValue followed by a ']'
-    *
-    * @return a Parser[Seq[Element]
-    */
-  def optionalElements: Parser[Seq[Element]] = openBracket ~> rep(optionWithOrWithoutValue) <~ closeBracket ^^ (ts => for (t <- ts) yield OptionalElement(t))
-
-  /**
-    * An optionWithOrWithoutValue matches EITHER: an command; OR: an command followed by a value
-    *
-    * @return a Parser[Element] which is EITHER: a Parser[Command] OR: a Parser[CommandWithValue]
-    */
-  def optionWithOrWithoutValue: Parser[Element] = (command ~ optionalValue | command ~ value | command) ^^ {
-    case o: Element => o
-    case (o: Element) ~ (v: Element) => CommandWithValue(o.value, v)
-    case _ => throw new Exception("")
-  }
-
-  /**
-    * An optionalValue matches "[" "value" "]"
-    *
-    * @return a Parser[OptionalElement]
-    */
-  def optionalValue: Parser[Element] = openBracket ~> value <~ closeBracket ^^ { e => OptionalElement(e) }
-
-  /**
-    * A value matches EITHER: a space [which is ignored] followed by a valueToken1 OR: a valueToken2
-    *
-    * @return a Parser[Value]
-    */
-  def value: Parser[Value] = ("""\s""".r ~> valueToken1 | valueToken2) ^^ (t => Value(t))
-
-  /**
-    * A command matches a single character which is either a lowercase letter or a digit
-    *
-    * @return a Parser[Command]
-    */
-  def command: Parser[Command] = """[\p{Ll}\d]""".r ^^ (t => Command(t))
-
-  /**
-    * A valueToken2 matches an uppercase letter followed by any number of non-space, non-bracket symbols
-    *
-    * @return a Parser[String]
-    */
-  val valueToken2: Parser[String] = """\p{Lu}[^\[\]\s]*""".r
-
-  /**
-    * A valueToken1 matches at least one non-space, non-bracket symbol
-    *
-    * @return a Parser[String]
-    */
-  val valueToken1: Parser[String] = """[^\[\]\s]+""".r
-
-  private val openBracket = """\[""".r
-  private val closeBracket = """\]""".r
-
-}
-
-abstract class ArgsException(s: String) extends Exception(s"Args exception: $s")
-
-case object AnonymousNotFoundException extends ArgsException("no anonymous arg found")
-
-case class NotFoundException(command: String) extends ArgsException(s"Arg: command $command not found")
-
-case class AmbiguousNameException(name: String) extends ArgsException(s"$name ambiguous")
-
-case class ParseException(cause: String) extends ArgsException(cause)
-
-case class NoValueException(name: Option[String]) extends ArgsException(s"Arg: command ${name.getOrElse("anonymous")} has no value")
-
-case class ValidationException[X](a: Args[X], s: Synopsis) extends ArgsException(s"Args: validation failed for $a with synopsis: $s")
-
-case class InvalidOptionException[X](arg: Arg[X]) extends ArgsException(s"Arg ${arg.name} not valid")
-
-case class CompareException(str: String) extends ArgsException(s"Arg compare exception: $str")
-
-case class NoOptionInSynopsisException(str: String) extends ArgsException(s"parse error: option $str was not found in synopsis")
