+++ src/main/scala/com/phasmidsoftware/args/Derivable.scala
@@ -1,0 +1,21 @@
+/*
+ * Copyright (c) 2018. Phasmid Software
+ */
+
+package com.phasmidsoftware.args
+
+/**
+  * Type-class trait to allow conversion from type X to type T
+  *
+  * @tparam T the result type
+  */
+trait Derivable[T] {
+  /**
+    * Method to convert an X to a T
+    *
+    * @param x the X value
+    * @tparam X the input type
+    * @return a T
+    */
+  def deriveFrom[X](x: X): T
+}
