+++ src/test/scala/com/phasmidsoftware/args/ArgsSpec.scala
@@ -200,175 +200,6 @@ class ArgsSpec extends FlatSpec with Matchers {
     filename shouldBe "argFilename"
   }
 
-  behavior of "SimpleArgParser"
-
-  val p = new SimpleArgParser
-
-  it should "parse command " + cmdF in {
-    p.parseAll(p.command,cmdF) should matchPattern { case p.Success(p.Command(`nameF`), _) => }
-  }
-
-  it should "not parse command -X" in {
-    p.parseAll(p.command,"-X") should matchPattern { case p.Failure(_, _) => }
-  }
-
-  it should "parse argument " + argFilename in {
-    p.parseAll(p.argument,argFilename) should matchPattern { case p.Success(p.Argument(`argFilename`), _) => }
-  }
-
-  it should "not parse argument -x" in {
-    p.parseAll(p.argument,"-x") should matchPattern { case p.Failure(_, _) => }
-  }
-
-  it should "parse token " + cmdF in {
-    p.parseAll(p.token,cmdF) should matchPattern { case p.Success(p.Command(`nameF`), _) => }
-  }
-
-  it should "parse token " + argFilename in {
-    p.parseAll(p.token,argFilename) should matchPattern { case p.Success(p.Argument(`argFilename`), _) => }
-  }
-
-  it should "parse " + cmdF in {
-    p.parseToken(cmdF) should matchPattern { case Success(p.Command(`nameF`)) => }
-  }
-
-  it should "parse " + argFilename in {
-    p.parseToken(argFilename) should matchPattern { case Success(p.Argument(`argFilename`)) => }
-  }
-
-  behavior of "PosixArgParser"
-
-  it should "parse options from -xf;" in {
-    val p = new PosixArgParser
-    val pp: p.ParseResult[PosixArg] = p.parseAll(p.posixOptions, "-xf;")
-    pp should matchPattern { case p.Success(_, _) => }
-  }
-
-  it should "parse operand from 3.1415927;" in {
-    val p = new PosixArgParser
-    val pp: p.ParseResult[PosixArg] = p.parseAll(p.posixOperand, "3.1415927;")
-    pp should matchPattern { case p.Success(_, _) => }
-  }
-
-  it should "parse option value from -argFilename;" in {
-    val p = new PosixArgParser
-    val pp: p.ParseResult[PosixArg] = p.parseAll(p.posixOptionValue, "argFilename;")
-    pp should matchPattern { case p.Success(_, _) => }
-  }
-
-  it should "parse option set from -xf;argFilename" in {
-    val p = new PosixArgParser
-    val pp: p.ParseResult[Seq[PosixArg]] = p.parseAll(p.posixOptionSet, "-xf;argFilename;")
-    pp should matchPattern { case p.Success(_, _) => }
-    pp.get.size shouldBe 2
-    pp.get.head shouldBe PosixOptionString("xf")
-    pp.get.last shouldBe PosixOptionValue("argFilename")
-  }
-
-  it should """parseCommandLine "-xf argFilename 3.1415927"""" in {
-    val p = new PosixArgParser
-    val as: Seq[PosixArg] = p.parseCommandLine(Seq("-xf", "argFilename", "3.1415927"))
-    as.size shouldBe 3
-    as.head shouldBe PosixOptionString("xf")
-    as.tail.head shouldBe PosixOptionValue("argFilename")
-    as.last shouldBe PosixOperand("3.1415927")
-  }
-
-  behavior of "PosixSynopsisParser"
-  it should "parse x as OptionToken" in {
-    val p = new PosixSynopsisParser
-    val cr = p.parse(p.command, "x")
-    cr should matchPattern { case p.Success(Command("x"), _) => }
-  }
-  it should "parse xf as two optionTokens" in {
-    val p = new PosixSynopsisParser
-    val eEr = p.parse(p.command ~ p.command, "xf")
-    eEr should matchPattern { case p.Success(_, _) => }
-  }
-  it should """parse "filename" as ValueToken1""" in {
-    val p = new PosixSynopsisParser
-    val wr = p.parse(p.valueToken1, "filename")
-    wr should matchPattern { case p.Success("filename", _) => }
-  }
-  it should "parse Junk as ValueToken2" in {
-    val p = new PosixSynopsisParser
-    val wr = p.parse(p.valueToken2, "Junk")
-    wr should matchPattern { case p.Success("Junk", _) => }
-  }
-  it should "parse Junk as ValueToken" in {
-    val p = new PosixSynopsisParser
-    val vr = p.parse(p.value, "Junk")
-    vr should matchPattern { case p.Success(Value("Junk"), _) => }
-  }
-  it should """parse " filename" as ValueToken""" in {
-    val p = new PosixSynopsisParser
-    val vr = p.parse(p.value, " filename")
-    vr should matchPattern { case p.Success(Value("filename"), _) => }
-  }
-  it should "parse x as an optionOrValueToken" in {
-    val p = new PosixSynopsisParser
-    val er = p.parse(p.optionWithOrWithoutValue, "x")
-    er should matchPattern { case p.Success(_, _) => }
-  }
-  it should "parse xf as two optionOrValueTokens" in {
-    val p = new PosixSynopsisParser
-    val eEr = p.parse(p.optionWithOrWithoutValue ~ p.optionWithOrWithoutValue, "xf")
-    eEr should matchPattern { case p.Success(_, _) => }
-  }
-  it should "parse [f] as an optionalElement" in {
-    val p = new PosixSynopsisParser
-    val er = p.parse(p.optionalElement, "[f]")
-    er should matchPattern { case p.Success(_, _) => }
-  }
-  it should "parse x[f] as one optionOrValueToken with three characters left over" in {
-    val p = new PosixSynopsisParser
-    val er = p.parse(p.optionWithOrWithoutValue, "x[f]")
-    er should matchPattern { case p.Success(Command("x"), _) => }
-    er.next.pos.column should matchPattern { case 2 => }
-  }
-  it should "parse x[f] as one optionOrValueToken followed by an optionalElement" in {
-    val p = new PosixSynopsisParser
-    val eEr = p.parse(p.optionWithOrWithoutValue ~ p.optionalElement, "x[f]")
-    eEr should matchPattern { case p.Success(_, _) => }
-  }
-  it should "parse x[f filename] as one optionOrValueToken followed by an optionalElement" in {
-    val p = new PosixSynopsisParser
-    val eEr = p.parse(p.optionWithOrWithoutValue ~ p.optionalElement, "x[f filename]")
-    eEr should matchPattern { case p.Success(_, _) => }
-  }
-  it should "parse f[ filename] as an optionalOrRequiredElement" in {
-    val p = new PosixSynopsisParser
-    val er = p.parse(p.optionalOrRequiredElement, "f[ filename]")
-    er should matchPattern { case p.Success(_, _) => }
-    er.get shouldBe CommandWithValue("f", OptionalElement(Value("filename")))
-  }
-  it should "parse f filename as a optionalOrRequiredElement" in {
-    val p = new PosixSynopsisParser
-    val esr = p.parse(p.optionalOrRequiredElement, "f filename")
-    esr should matchPattern { case p.Success(_, _) => }
-  }
-  it should "parse [xf filename] as a List[Element]" in {
-    val p = new PosixSynopsisParser
-    val esr = p.parse(p.optionalElements, "[xf filename]")
-    esr should matchPattern { case p.Success(_, _) => }
-  }
-  it should "parse -x[f filename] as a synopsis" in {
-    val p = new PosixSynopsisParser
-    val so: Option[Synopsis] = p.parseSynopsis(Some("-x[f filename]"))
-    so shouldBe Some(Synopsis(Seq(Command("x"), OptionalElement(CommandWithValue("f", Value("filename"))))))
-  }
-  it should "parse -[xf filename]" in {
-    val p = new PosixSynopsisParser
-    val so: Option[Synopsis] = p.parseSynopsis(Some("-[xf filename]"))
-    so shouldBe Some(Synopsis(Seq(OptionalElement(Command("x")), OptionalElement(CommandWithValue("f", Value("filename"))))))
-  }
-  it should "parse -x[f[ filename]]" in {
-    val p = new PosixSynopsisParser
-    val so: Option[Synopsis] = p.parseSynopsis(Some("-x[f[ filename]]"))
-    so shouldBe Some(Synopsis(Seq(Command("x"), OptionalElement(CommandWithValue("f", Value("filename"))))))
-  }
-
-
   behavior of "Args validation"
   it should "parse " + cmdF + " " + argFilename in {
     val args = Array(cmdF, argFilename, "positionalArg")
